refactored files into a nicer tree
