+++ index.html
@@ -32,7 +32,7 @@
         <div class='meal-foods'></div>
         <div class='row'>
           <div class='d-head-name'>Total Calories</div>
-          <div id='total-cals' class='d-head-cals'>0</div>
+          <div id='total-cals' class='d-head-cals'></div>
         </div>
         <div class='row'>
           <div class='d-head-name'>Remaining Calories</div>
